Server now sends pings every X milliseconds; entities can now be removed properly; improved sorting "algorithm"
Server now pings the client with entities based on time, not every single game tick. Still kinda laggy/stuttery but not nearly as bad as before. Also apparently too much packet loss causes your character to spontaneously stop existing when boosting too far out. The issue of entities not being removed is now fixed and I fucking hate myself for not realizing the issue was the break command preventing it from working (for SOME reason...). The sorting "algorithm" (if you can call it that...) is now 3*O(x) opposed to 4*O(x).