MASSIVE (internal) HUD REWORK + more HUD types added
YOU CAN NOW DISPLAY IMAGES ON THE HUD & THE HUD IS NOW MORE SIMPLE TO WORK WITH. I DID NOT KNOW I COULD DO THIS THE ENTIRE TIME BUT APPARENTLY I DIDNT EVEN HAVE TO ADD ANYTHING OR CHANGE THE WAY I IMPLEMENTED THE HUD TO GET IT TO WORK.
Additionally, I added the ability to format strings in the HUD class; It basically just adds a newline in place of a space between two specific words, determined by the length. Also added a new model, changed the size & distance of the model used alongside a new placeholder photo because why not.